Conflicts:
	libcflie_delivery/src/control/control.cpp

+++ libcflie_delivery/include/cflie/CCrazyflie.h
@@ -150,7 +150,7 @@ typedef struct _CCrazyflie {
     \param nThrust The thrust value to send (> 10000) */
   void setThrust(CCrazyflie* crFile,int nThrust);
   
-  void setYaw(CCrazyflie* crFile,int nYaw);
+  void setYaw(CCrazyflie* crFile, float nYaw);
 
 
   /*! \brief Set the roll control set point
